+++ App.js
@@ -2,31 +2,11 @@ import { AppLoading } from "expo";
 import { Asset } from "expo-asset";
 import * as Font from "expo-font";
 import React, { useState } from "react";
-import { Platform, StatusBar, StyleSheet, View } from "react-native";
+import { Platform, StatusBar, StyleSheet, View, TouchableOpacity, Text } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import AppNavigator from "./navigation/AppNavigator";
-import * as firebase from "firebase";
-import {
-  API_KEY,
-  AUTH_DOMAIN,
-  DATABASE_URL,
-  PROJECT_ID,
-  MESSAGE_SENDER_ID,
-  APP_ID
-} from "react-native-dotenv";
-
-// Initialize Firebase
-const firebaseConfig = {
-  apiKey: API_KEY,
-  authDomain: AUTH_DOMAIN,
-  databaseURL: DATABASE_URL,
-  projectId: PROJECT_ID,
-  storageBucket: "",
-  messagingSenderId: MESSAGE_SENDER_ID,
-  appId: APP_ID
-};
-
-firebase.initializeApp(firebaseConfig);
+import firebase from "./utils/firebase/firebase";
+import signInWithFacebook from "./utils/auth/auth";
 
 function storeHighScore(userId, score) {
   firebase
@@ -55,6 +35,13 @@ export default function App(props) {
       <View style={styles.container}>
         {Platform.OS === "ios" && <StatusBar barStyle="default" />}
         <AppNavigator />
+        <View>
+          <TouchableOpacity onPress={handleTouch}>
+            <Text>
+              handleTouch
+            </Text>
+          </TouchableOpacity>
+        </View>
       </View>
     );
   }
@@ -84,6 +71,10 @@ function handleLoadingError(error) {
 
 function handleFinishLoading(setLoadingComplete) {
   setLoadingComplete(true);
+}
+
+function handleTouch() {
+  signInWithFacebook()
 }
 
 const styles = StyleSheet.create({
