+++ package.json
@@ -18,6 +18,7 @@
     "expo": "^35.0.0",
     "expo-asset": "^7.0.0",
     "expo-constants": "^7.0.0",
+    "expo-facebook": "~7.0.0",
     "expo-font": "^7.0.0",
     "expo-web-browser": "^7.0.0",
     "firebase": "^7.2.1",
